Split out shape-changing Time methods to mixin class.
